+++ src/pages/@owner/estimate/@section/basic-info-section/index.tsx
@@ -1,4 +1,4 @@
-export { default as Menu } from '@pages/@owner/estimate/@section/basic-info-section/Menu';
+export { default as Food } from '@pages/@owner/estimate/@section/basic-info-section/Food';
 export { default as Price } from '@pages/@owner/estimate/@section/basic-info-section/Price';
 export { default as NeedElectricity } from '@pages/@owner/estimate/@section/basic-info-section/NeedElectricity';
 export { default as Etc } from '@pages/@owner/estimate/@section/basic-info-section/Etc';
