+++ src/pages/@owner/estimate/utils/estimate.schema.ts
@@ -26,14 +26,23 @@ export const estimateSchema = z.object({
       ESTIMATE_MAX_LENGTH.availableDates.max,
       ESTIMATE_ERROR_MESSAGE.availableDates.max
     ),
-  activeTime: z.object({
-    startActiveTime: z.string(),
-    endActiveTime: z.string(),
-  }),
-  menu: z
+  activeTime: z
+    .object({
+      startActiveTime: z.string(),
+      endActiveTime: z.string(),
+    })
+    .refine(
+      data => {
+        return data.startActiveTime < data.endActiveTime;
+      },
+      {
+        message: ESTIMATE_ERROR_MESSAGE.activeTime.invalid,
+      }
+    ),
+  food: z
     .string()
-    .min(ESTIMATE_MAX_LENGTH.menu.min, ESTIMATE_ERROR_MESSAGE.menu.required)
-    .max(ESTIMATE_MAX_LENGTH.menu.max, ESTIMATE_ERROR_MESSAGE.menu.max),
+    .min(ESTIMATE_MAX_LENGTH.food.min, ESTIMATE_ERROR_MESSAGE.food.required)
+    .max(ESTIMATE_MAX_LENGTH.food.max, ESTIMATE_ERROR_MESSAGE.food.max),
   price: z
     .number()
     .refine(val => val > 0, ESTIMATE_ERROR_MESSAGE.price.required),
